[Refactor] Refactor CCE nodepools list
