--- openstack/cce/v3/nodepools/urls.go
@@ -1,12 +1,0 @@
-package nodepools
-
-import golangsdk "github.com/opentelekomcloud/gophertelekomcloud"
-
-const (
-	rootPath     = "clusters"
-	resourcePath = "nodepools"
-)
-
-func rootURL(c *golangsdk.ServiceClient, clusterid string) string {
-	return c.ServiceURL(rootPath, clusterid, resourcePath)
-}
